+++ examples/avif_example_encode.c
@@ -27,7 +27,7 @@ int main(int argc, char * argv[])
 
     avifImage * image = avifImageCreate(128, 128, 8, AVIF_PIXEL_FORMAT_YUV444); // these values dictate what goes into the final AVIF
     if (!image) {
-        fprintf(stderr, "Out-of-memory\n");
+        fprintf(stderr, "Out of memory\n");
         goto cleanup;
     }
     // Configure image here: (see avif/avif.h)
@@ -80,7 +80,7 @@ int main(int argc, char * argv[])
 
     encoder = avifEncoderCreate();
     if (!encoder) {
-        fprintf(stderr, "Out-of-memory\n");
+        fprintf(stderr, "Out of memory\n");
         goto cleanup;
     }
     // Configure your encoder here (see avif/avif.h):